removed Nsb Project and added Nuget Package instead
after debugging the actual NSB implementation and extending it with our custom needs
we can go back to referencing NSB via NuGet to see it in action

As a side note, the Azure Transport doesn't look to be very extendable atm.
Hopefully there will be some new releases soon, to make it a bit more flexible.